Cinema/CinemaApp/requirements.txt
flask
google-cloud-firestore
